+++ src/Core/Plugin.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
  */
 namespace MeTools\Core;
 
-use Cake\Core\Exception\MissingPluginException;
 use Cake\Core\Plugin as CakePlugin;
+use Tools\Exceptionist;
 use Tools\Filesystem;
 
 /**
@@ -58,7 +58,7 @@ class Plugin extends CakePlugin
      * @param string|null $file File
      * @param bool $check Checks if the file exists
      * @return string Path of the plugin or path of the path of a plugin file
-     * @throws \ErrorException
+     * @throws \Tools\Exception\FileNotExistsException|\ErrorException
      */
     public static function path(string $name, ?string $file = null, bool $check = false): string
     {
@@ -68,8 +68,8 @@ class Plugin extends CakePlugin
         }
 
         $path = $plugin . $file;
-        if (!is_readable($path) && $check) {
-            throw new MissingPluginException(__d('me_tools', 'File or directory `{0}` does not exist', Filesystem::instance()->rtr($path)));
+        if ($check) {
+            Exceptionist::fileExists($path, __d('me_tools', 'File or directory `{0}.{1}` does not exist', $name, Filesystem::instance()->rtr($path)));
         }
 
         return $path;